+++ NEWS.md
@@ -1,3 +1,5 @@
+# medsim (development version)
+
 # medsim 0.5.0
 
 ## New features