+++ NEWS.md
@@ -1,4 +1,4 @@
-# medsim (development version)
+# medsim 0.5.1
 
 ## New features